% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Telemetry_module_API.R
\name{VividTelemetry}
\alias{VividTelemetry}
\title{VividTelemetry}
\description{
A telemetry module for the Vivid Volcano app that tracks user activity
using Supabase REST API with SQLite fallback.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VividTelemetry-new}{\code{VividTelemetry$new()}}
\item \href{#method-VividTelemetry-parse_user_agent}{\code{VividTelemetry$parse_user_agent()}}
\item \href{#method-VividTelemetry-setup_sqlite_database}{\code{VividTelemetry$setup_sqlite_database()}}
\item \href{#method-VividTelemetry-update_visitor_info}{\code{VividTelemetry$update_visitor_info()}}
\item \href{#method-VividTelemetry-update_sqlite_visitor_info}{\code{VividTelemetry$update_sqlite_visitor_info()}}
\item \href{#method-VividTelemetry-increment_counter}{\code{VividTelemetry$increment_counter()}}
\item \href{#method-VividTelemetry-end_session}{\code{VividTelemetry$end_session()}}
\item \href{#method-VividTelemetry-end_session_sqlite}{\code{VividTelemetry$end_session_sqlite()}}
\item \href{#method-VividTelemetry-get_sqlite_connection}{\code{VividTelemetry$get_sqlite_connection()}}
\item \href{#method-VividTelemetry-print_status}{\code{VividTelemetry$print_status()}}
\item \href{#method-VividTelemetry-clone}{\code{VividTelemetry$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VividTelemetry-new"></a>}}
\if{latex}{\out{\hypertarget{method-VividTelemetry-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VividTelemetry$new(user_agent = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VividTelemetry-parse_user_agent"></a>}}
\if{latex}{\out{\hypertarget{method-VividTelemetry-parse_user_agent}{}}}
\subsection{Method \code{parse_user_agent()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VividTelemetry$parse_user_agent(user_agent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VividTelemetry-setup_sqlite_database"></a>}}
\if{latex}{\out{\hypertarget{method-VividTelemetry-setup_sqlite_database}{}}}
\subsection{Method \code{setup_sqlite_database()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VividTelemetry$setup_sqlite_database()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VividTelemetry-update_visitor_info"></a>}}
\if{latex}{\out{\hypertarget{method-VividTelemetry-update_visitor_info}{}}}
\subsection{Method \code{update_visitor_info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VividTelemetry$update_visitor_info(visit_count)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VividTelemetry-update_sqlite_visitor_info"></a>}}
\if{latex}{\out{\hypertarget{method-VividTelemetry-update_sqlite_visitor_info}{}}}
\subsection{Method \code{update_sqlite_visitor_info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VividTelemetry$update_sqlite_visitor_info()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VividTelemetry-increment_counter"></a>}}
\if{latex}{\out{\hypertarget{method-VividTelemetry-increment_counter}{}}}
\subsection{Method \code{increment_counter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VividTelemetry$increment_counter(counter_type)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VividTelemetry-end_session"></a>}}
\if{latex}{\out{\hypertarget{method-VividTelemetry-end_session}{}}}
\subsection{Method \code{end_session()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VividTelemetry$end_session()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VividTelemetry-end_session_sqlite"></a>}}
\if{latex}{\out{\hypertarget{method-VividTelemetry-end_session_sqlite}{}}}
\subsection{Method \code{end_session_sqlite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VividTelemetry$end_session_sqlite()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VividTelemetry-get_sqlite_connection"></a>}}
\if{latex}{\out{\hypertarget{method-VividTelemetry-get_sqlite_connection}{}}}
\subsection{Method \code{get_sqlite_connection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VividTelemetry$get_sqlite_connection()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VividTelemetry-print_status"></a>}}
\if{latex}{\out{\hypertarget{method-VividTelemetry-print_status}{}}}
\subsection{Method \code{print_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VividTelemetry$print_status()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VividTelemetry-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VividTelemetry-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VividTelemetry$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
